steel/requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
# Install numpy before torch — torch links against numpy at install time
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2

# ── Vision / Detection ────────────────────────────────────────────────────────
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# ── App / UI ──────────────────────────────────────────────────────────────────
streamlit==1.32.2
plotly==5.20.0
pandas==2.2.2

# ── Firebase ──────────────────────────────────────────────────────────────────
firebase-admin==6.4.0
